Added emoji library, basic parser support
Added the emoji library from EmojiOne. Removed the "512 x 512" images because they really increased the size of the output DLL. (It's still 5 MB, but whatever). Also added a EmojiBoxSample project with a basic window to test out the library on! All buttons in the window work. (We have a working emoji name parser, woo!)